refactor: split GC reporting and resource helpers
