+++ packages/stim-cli/src/commands/gc.ts
@@ -1,128 +1,59 @@
-import { existsSync, readFileSync, readdirSync, realpathSync, rmSync, statSync } from 'fs';
-import { homedir, tmpdir } from 'os';
-import { basename, dirname, isAbsolute, join, relative, resolve } from 'path';
+import { existsSync, rmSync, statSync } from 'fs';
 import chalk from 'chalk';
 import { InvalidArgumentError, type Command } from 'commander';
-import { clearDevice, getConfigDir, loadConfig } from '../config.ts';
-import { formatLongDuration, plural, shortUdid } from '../command-output.ts';
-import { directorySize, formatBytes, isOnMountedVolume, listMountedVolumes, volumeRootFor } from '../fs-util.ts';
+import { loadConfig } from '../config.ts';
+import { directorySize, isOnMountedVolume, listMountedVolumes, volumeRootFor } from '../fs-util.ts';
 import { listBuildLocks, readBuildLock } from '../engine/build-lock.ts';
 import { listBuildSlots, readBuildSlot } from '../engine/build-slots.ts';
-import { leaseIsExpired, listLeaseFiles, removeExpiredLease, type LeaseFileEntry } from '../engine/device-lease.ts';
-import {
-  findOrphanedOwnedSessions,
-  getSessionArgs,
-  inspectSessionForTeardown,
-  isDefinitiveMissingSessionError,
-  listOwnedSessionsArgs,
-  parseSessionListPage,
-  stopSessionArgs,
-  verifyStoppedSession,
-  type ScopedSessionSummary,
-} from '../engine/eas-simulator.ts';
-import { resolveEasCliBin } from '../engine/remote-cache.ts';
-import { getExecutor } from '../exec.ts';
+import { removeExpiredLease } from '../engine/device-lease.ts';
 import { isPidAlive } from '../metro.ts';
 import { detectIsExpo, findProjectRoot } from '../project.ts';
 import { reclaimProject } from '../reclaim.ts';
 import { SETTING_SHAPE_REMEDY } from '../settings.ts';
-import { listAllIosSims, listIosDeviceTypes, parseRuntimeVersion, type IosSimRecord } from '../sim/ios.ts';
-import { dropParked, parkedMaxSetting, POOL_SETTING_REMEDY, readParked, type ParkedSim } from '../sim-pool.ts';
-import { teardownOwnedIosSim, teardownOwnedAvd, teardownParkedIosSim } from '../teardown.ts';
+import { listAllIosSims, type IosSimRecord } from '../sim/ios.ts';
+import { parkedMaxSetting, POOL_SETTING_REMEDY } from '../sim-pool.ts';
 import { listAvds, ownedAvdDirectory } from '../sim/android.ts';
-import {
-  declaredCachePaths,
-  discoverCaches,
-  projectSettingShapeErrors,
-  pruneCache,
-  sizeCaches,
-  type CacheDescriptor,
-} from '../caches.ts';
-import { workspaceDir, workspaceStateFile } from '../paths.ts';
-import { withRemoteSessionLock } from '../engine/device-remote.ts';
+import { declaredCachePaths, discoverCaches, projectSettingShapeErrors, sizeCaches } from '../caches.ts';
 import { withEasProjectLock } from '../engine/eas-project-lock.ts';
+import type { GcSkip, OrphanedDevice } from '../types.ts';
+import { emptyCaches, planCacheEmptying, selectCaches, trimCaches } from './gc/caches.ts';
 import {
-  easMachineStateRoot,
-  readEasSessionLedger,
-  removeEasSessionClaim,
-  type EasSessionClaim,
-} from '../engine/eas-session-ledger.ts';
-import type { BuildLockInfo, BuildSlotInfo, Config, GcSkip, OrphanedDevice } from '../types.ts';
+  collectDeviceLeases,
+  collectParkedSims,
+  deleteParkedSims,
+  deleteProjectDevices,
+  describeUnverifiableDevices,
+  deviceSweepIsScoped,
+  DEVICE_LIST_TIMEOUT_MS,
+  findOrphanedDevices,
+  findStaleDeviceRecords,
+  findStaleProjectDevices,
+  withAndroidAvdSizes,
+  type GcDeviceDependencies,
+  type StaleDeviceRecord,
+  type StaleProjectDevice,
+} from './gc/devices.ts';
+import {
+  collectEasSessionSweep,
+  deleteEasSessions,
+  describeError,
+  withRemoteSessionGcLocks,
+  type EasGcDependencies,
+  type EasSessionSweep,
+} from './gc/eas-sessions.ts';
+import { formatGcReport, type GcReport } from './gc/report.ts';
 
-interface StaleProjectDevice {
-  kind: 'ios' | 'android';
-  id: string;
-  name: string;
-  project: string;
-  idleDays: number;
-  bytes?: number;
-}
-
-interface StaleDeviceRecord {
-  kind: 'ios' | 'android';
-  id: string;
-  project: string;
-  owned: boolean;
-}
-
-interface KeptDevice {
-  kind: 'ios' | 'android';
-  id: string;
-  name: string;
-  reason: string;
-}
-
-interface GcCache extends CacheDescriptor {
-  machineGlobal?: string | null;
-  willEmpty?: boolean;
-  emptySkipped?: string | null;
-}
-
-interface EasSessionSweep {
-  projectScope: string | null;
-  orphaned: ScopedSessionSummary[];
-  notices: string[];
-  deletionSafe: boolean;
-}
-
-interface KeptLeaseFile {
-  name: string;
-  path: string;
-  reason: string;
-}
-
-interface DeviceLeaseGarbage {
-  expired: LeaseFileEntry[];
-  kept: KeptLeaseFile[];
-}
-
-export interface ParkedSimReport {
-  udid: string;
-  name: string;
-  model: string | null;
-  runtime: string | null;
-  parkedAt: string;
-  bytes: number | null;
-  listed: boolean | null;
-}
-
-interface GcReport {
-  skipped: GcSkip[];
-  deadProjects: string[];
-  parkedSims: ParkedSimReport[];
-  orphanedDevices: OrphanedDevice[];
-  staleDevices: StaleProjectDevice[];
-  staleDeviceRecords: StaleDeviceRecord[];
-  buildLocks: { stale: BuildLockInfo[]; live: BuildLockInfo[] };
-  buildSlots: { stale: BuildSlotInfo[]; live: BuildSlotInfo[] };
-  deviceLeases: DeviceLeaseGarbage;
-  deviceSweepNotices: string[];
-  easSessionSweep: EasSessionSweep;
-  caches: GcCache[];
-  cacheScope: string | null;
-  olderThan: number | null;
-  all: boolean;
-}
+export { selectCaches } from './gc/caches.ts';
+export {
+  deleteParkedSims,
+  describeParkedSims,
+  describeUnverifiableDevices,
+  findOrphanedDevices,
+  findStaleDeviceRecords,
+  findStaleProjectDevices,
+  type ParkedSimReport,
+} from './gc/devices.ts';
+export { formatGcReport } from './gc/report.ts';
 
 interface CollectGcReportOptions {
   olderThan?: number | null;
@@ -139,802 +70,8 @@ interface RunGcOptions {
   unsafeAllowScopedDeviceSweep?: boolean;
 }
 
-interface GcDependencies {
-  findProjectRoot?: typeof findProjectRoot;
-  detectIsExpo?: typeof detectIsExpo;
-  resolveEasCliBin?: typeof resolveEasCliBin;
-  runEasFile?: (
-    file: string,
-    args: string[],
-    options: { cwd: string; timeoutMs: number; omitEnv: readonly string[] },
-  ) => string;
-  withRemoteSessionLock?: typeof withRemoteSessionLock;
-  withEasProjectLock?: typeof withEasProjectLock;
-  easMaxPages?: number;
-  easCollectionTimeoutMs?: number;
-  easNow?: () => number;
-  easLockWaitMs?: number;
-  easLockSnapshotRetries?: number;
-  easSweepBlockedNotice?: string;
-  easLockedRoots?: readonly string[];
-  easLedgerRoot?: string;
-  removeEasSessionClaim?: typeof removeEasSessionClaim;
-  precollectedEasSessionSweep?: EasSessionSweep;
-  avdDirectory?: typeof ownedAvdDirectory;
-  directorySize?: typeof directorySize;
+interface GcDependencies extends EasGcDependencies, GcDeviceDependencies {
   settingShapeErrors?: () => string[];
-  listAllIosSims?: typeof listAllIosSims;
-  listIosDeviceTypes?: typeof listIosDeviceTypes;
-  deleteParkedIosSim?: (udid: string) => void;
-}
-
-// simctl and emulator listings can exceed 10 seconds on loaded hosts; 30 seconds still bounds hangs.
-const DEVICE_LIST_TIMEOUT_MS = 30000;
-const AVD_SIZE_TIMEOUT_MS = 5000;
-const EAS_OPERATION_TIMEOUT_MS = 30000;
-const EAS_COLLECTION_TIMEOUT_MS = 60000;
-const EAS_MAX_LIST_PAGES = 100;
-const EAS_LOCK_WAIT_MS = 0;
-const EAS_LOCK_SNAPSHOT_RETRIES = 3;
-const PROXY_CREDENTIAL_ENV = ['AGENT_DEVICE_DAEMON_BASE_URL', 'AGENT_DEVICE_DAEMON_AUTH_TOKEN'] as const;
-
-const DAY_MS = 24 * 60 * 60 * 1000;
-
-function describeKept(ref: { path: string; mounted: boolean }): string {
-  return ref.mounted
-    ? `referenced by ${ref.path}`
-    : `referenced by ${ref.path} (volume not mounted; kept just in case)`;
-}
-
-export function findOrphanedDevices({
-  sims = [],
-  avds = [],
-  config,
-  isMounted,
-  deadProjects = [],
-}: {
-  sims?: IosSimRecord[];
-  avds?: string[];
-  config: Config | null;
-  isMounted?: (path: string) => boolean;
-  deadProjects?: string[];
-}): { orphaned: OrphanedDevice[]; kept: KeptDevice[] } {
-  const dead = new Set(deadProjects);
-  const referenced = new Map<string, { path: string; mounted: boolean }>();
-
-  for (const [path, proj] of Object.entries(config?.projects || {})) {
-    if (dead.has(path)) continue;
-    const mounted = isMounted ? isMounted(path) : true;
-    const ios = proj?.platforms?.ios;
-    if (ios?.deviceUdid) {
-      referenced.set(ios.deviceUdid, { path, mounted });
-    }
-    const android = proj?.platforms?.android;
-    if (android?.avdName) {
-      referenced.set(android.avdName, { path, mounted });
-    }
-  }
-  for (const sim of readParked('ios', { config })) {
-    referenced.set(sim.udid, { path: 'the simulator pool', mounted: true });
-  }
-
-  const orphaned: OrphanedDevice[] = [];
-  const kept: KeptDevice[] = [];
-
-  for (const sim of sims) {
-    if (!sim?.name?.startsWith('stim-')) continue;
-    const ref = referenced.get(sim.udid);
-    if (!ref) {
-      orphaned.push({ kind: 'ios', id: sim.udid, name: sim.name });
-    } else {
-      kept.push({ kind: 'ios', id: sim.udid, name: sim.name, reason: describeKept(ref) });
-    }
-  }
-
-  for (const avdName of avds) {
-    if (!avdName?.startsWith('stim-')) continue;
-    const ref = referenced.get(avdName);
-    if (!ref) {
-      orphaned.push({ kind: 'android', id: avdName, name: avdName });
-    } else {
-      kept.push({ kind: 'android', id: avdName, name: avdName, reason: describeKept(ref) });
-    }
-  }
-
-  return { orphaned, kept };
-}
-
-export function findStaleProjectDevices({
-  config,
-  sims = [],
-  avds = [],
-  olderThanDays,
-  now = Date.now(),
-  lastTouched,
-  deadProjects = [],
-}: {
-  config: Config | null;
-  sims?: IosSimRecord[];
-  avds?: string[];
-  olderThanDays?: number;
-  now?: number;
-  lastTouched?: (path: string) => number;
-  deadProjects?: string[];
-}): StaleProjectDevice[] {
-  if (!Number.isFinite(olderThanDays) || typeof lastTouched !== 'function') return [];
-  const cutoff = now - (olderThanDays as number) * DAY_MS;
-  const dead = new Set(deadProjects);
-
-  const liveSims = new Map<string, string>(
-    sims.filter((s) => s?.name?.startsWith('stim-')).map((s) => [s.udid, s.name] as [string, string]),
-  );
-  const liveAvds = new Set(avds.filter((a) => typeof a === 'string' && a.startsWith('stim-')));
-
-  const stale: StaleProjectDevice[] = [];
-  for (const [path, proj] of Object.entries(config?.projects || {})) {
-    if (dead.has(path)) continue;
-    const touched = lastTouched(path);
-    if (!Number.isFinite(touched) || touched >= cutoff) continue;
-    const idleDays = Math.floor((now - touched) / DAY_MS);
-
-    const ios = proj?.platforms?.ios;
-    if (ios?.owned && ios.deviceUdid && liveSims.has(ios.deviceUdid)) {
-      stale.push({
-        kind: 'ios',
-        id: ios.deviceUdid,
-        name: liveSims.get(ios.deviceUdid) as string,
-        project: path,
-        idleDays,
-      });
-    }
-    const android = proj?.platforms?.android;
-    if (android?.owned && android.avdName && liveAvds.has(android.avdName)) {
-      stale.push({ kind: 'android', id: android.avdName, name: android.avdName, project: path, idleDays });
-    }
-  }
-  return stale;
-}
-
-export function findStaleDeviceRecords({
-  config,
-  sims = [],
-  avds = [],
-  deadProjects = [],
-  simsChecked = true,
-  avdsChecked = true,
-}: {
-  config: Config | null;
-  sims?: IosSimRecord[];
-  avds?: string[];
-  deadProjects?: string[];
-  simsChecked?: boolean;
-  avdsChecked?: boolean;
-}): StaleDeviceRecord[] {
-  const dead = new Set(deadProjects);
-  const liveSims = new Set(sims.map((s) => s?.udid).filter(Boolean));
-  const liveAvds = new Set(avds.filter((a) => typeof a === 'string'));
-
-  const stale: StaleDeviceRecord[] = [];
-  for (const [path, proj] of Object.entries(config?.projects || {})) {
-    if (dead.has(path)) continue;
-
-    const ios = proj?.platforms?.ios;
-    if (simsChecked && ios?.deviceUdid && !liveSims.has(ios.deviceUdid)) {
-      stale.push({ kind: 'ios', id: ios.deviceUdid, project: path, owned: Boolean(ios.owned) });
-    }
-    const android = proj?.platforms?.android;
-    if (avdsChecked && android?.avdName && !liveAvds.has(android.avdName)) {
-      stale.push({ kind: 'android', id: android.avdName, project: path, owned: Boolean(android.owned) });
-    }
-  }
-  return stale;
-}
-
-export function describeUnverifiableDevices(
-  simNames: string[] = [],
-  avdNames: string[] = [],
-  { reason = 'no Stim config found' }: { reason?: string } = {},
-): string[] {
-  const ours = [...simNames, ...avdNames].filter((n) => typeof n === 'string' && n.startsWith('stim-'));
-  if (ours.length === 0) return [`${reason}; device sweep skipped`];
-  return [
-    `${reason}, so ${ours.length} stim-created device(s) cannot be verified as orphaned: ${ours.join(', ')}`,
-    'they were NOT touched. If they are stale, delete them with `xcrun simctl delete <udid>` or `avdmanager delete avd -n <name>`',
-  ];
-}
-
-function describeError(error: unknown): string {
-  const candidate = error as { stderr?: unknown; message?: unknown };
-  if (typeof candidate?.stderr === 'string' && candidate.stderr.trim()) return candidate.stderr.trim();
-  if (typeof candidate?.message === 'string' && candidate.message.trim()) return candidate.message.trim();
-  return String(error);
-}
-
-function readRecordedRemoteSessionIds(roots: string[]): {
-  ids: string[];
-  notices: string[];
-  safe: boolean;
-} {
-  const ids = new Set<string>();
-  const notices: string[] = [];
-  let safe = true;
-
-  for (const root of new Set(roots)) {
-    try {
-      if (!statSync(root).isDirectory()) throw new Error('path is not a directory');
-      readdirSync(root);
-    } catch (error) {
-      safe = false;
-      notices.push(`${root} is not available as a readable workspace directory: ${describeError(error)}`);
-      continue;
-    }
-
-    const path = workspaceStateFile(root);
-    if (!existsSync(path)) continue;
-    let state: unknown;
-    try {
-      state = JSON.parse(readFileSync(path, 'utf-8')) as unknown;
-    } catch (error) {
-      safe = false;
-      notices.push(`${path} could not be read as valid JSON: ${describeError(error)}`);
-      continue;
-    }
-    if (!state || typeof state !== 'object' || Array.isArray(state)) {
-      safe = false;
-      notices.push(`${path} does not contain a valid workspace state object.`);
-      continue;
-    }
-    const remote = (state as { remoteDevice?: unknown }).remoteDevice;
-    if (remote === undefined) continue;
-    if (!remote || typeof remote !== 'object' || Array.isArray(remote)) {
-      safe = false;
-      notices.push(`${path} has a malformed remote session record.`);
-      continue;
-    }
-    const id = (remote as { sessionId?: unknown }).sessionId;
-    if (typeof id !== 'string' || id.length === 0) {
-      safe = false;
-      notices.push(`${path} has a remote session record with no valid session id.`);
-      continue;
-    }
-    ids.add(id);
-  }
-  return { ids: [...ids], notices, safe };
-}
-
-function claimStateAbsence(claim: EasSessionClaim): { absent: boolean; notice: string | null } {
-  const stateDir = dirname(claim.stateFile);
-  let entries: string[];
-  try {
-    if (!statSync(stateDir).isDirectory()) throw new Error('path is not a directory');
-    entries = readdirSync(stateDir);
-  } catch (error) {
-    return {
-      absent: false,
-      notice: `${stateDir} is not available for EAS ownership verification: ${describeError(error)}`,
-    };
-  }
-  if (!entries.includes(basename(claim.stateFile))) return { absent: true, notice: null };
-  let state: unknown;
-  try {
-    state = JSON.parse(readFileSync(claim.stateFile, 'utf-8')) as unknown;
-  } catch (error) {
-    return { absent: false, notice: `${claim.stateFile} could not be read as valid JSON: ${describeError(error)}` };
-  }
-  if (!state || typeof state !== 'object' || Array.isArray(state)) {
-    return { absent: false, notice: `${claim.stateFile} does not contain a valid workspace state object.` };
-  }
-  const remote = (state as { remoteDevice?: unknown }).remoteDevice;
-  if (remote === undefined) return { absent: true, notice: null };
-  if (!remote || typeof remote !== 'object' || Array.isArray(remote)) {
-    return { absent: false, notice: `${claim.stateFile} has a malformed remote session record.` };
-  }
-  const sessionId = (remote as { sessionId?: unknown }).sessionId;
-  if (typeof sessionId !== 'string' || !sessionId) {
-    return { absent: false, notice: `${claim.stateFile} has a remote session record with no valid session id.` };
-  }
-  return { absent: sessionId !== claim.sessionId, notice: null };
-}
-
-function collectEasSessionSweep(config: Config | null, deps: GcDependencies): EasSessionSweep {
-  const projectRoot = (deps.findProjectRoot ?? findProjectRoot)(process.cwd());
-  const empty = (notices: string[] = []): EasSessionSweep => ({
-    projectScope: projectRoot,
-    orphaned: [],
-    notices,
-    deletionSafe: notices.length === 0,
-  });
-  if (!projectRoot || !(deps.detectIsExpo ?? detectIsExpo)(projectRoot)) return empty();
-  if (deps.easSweepBlockedNotice) return empty([deps.easSweepBlockedNotice]);
-
-  const eas = (deps.resolveEasCliBin ?? resolveEasCliBin)(projectRoot);
-  if (!eas) return empty(['EAS session sweep skipped: eas-cli is not available for the current Expo project.']);
-
-  const workspaceRoots = [...Object.keys(config?.projects ?? {}), projectRoot];
-  const recorded = readRecordedRemoteSessionIds(workspaceRoots);
-  const ledgerRoot = deps.easLedgerRoot ?? easMachineStateRoot();
-  const ledger = readEasSessionLedger(ledgerRoot);
-  const run = deps.runEasFile ?? ((file, args, options) => getExecutor().runFile(file, args, options));
-  const sessions: unknown[] = [];
-  const seenCursors = new Set<string>();
-  const now = deps.easNow ?? Date.now;
-  const collectionTimeoutMs = deps.easCollectionTimeoutMs ?? EAS_COLLECTION_TIMEOUT_MS;
-  const maxPages = deps.easMaxPages ?? EAS_MAX_LIST_PAGES;
-  const deadline = now() + collectionTimeoutMs;
-  let pageCount = 0;
-  let after: string | null = null;
-  while (true) {
-    if (pageCount >= maxPages) {
-      return empty([
-        ...recorded.notices,
-        `EAS session list exceeded the maximum page limit of ${maxPages} for ${projectRoot}.`,
-      ]);
-    }
-    const remainingMs = deadline - now();
-    if (remainingMs <= 0) {
-      return empty([...recorded.notices, `EAS session list exceeded its total time limit for ${projectRoot}.`]);
-    }
-
-    let stdout: string;
-    try {
-      stdout = run(eas.file, listOwnedSessionsArgs(after), {
-        cwd: projectRoot,
-        timeoutMs: Math.min(EAS_OPERATION_TIMEOUT_MS, remainingMs),
-        omitEnv: PROXY_CREDENTIAL_ENV,
-      });
-    } catch (error) {
-      return empty([...recorded.notices, `EAS session list failed for ${projectRoot}: ${describeError(error)}`]);
-    }
-    pageCount++;
-    if (now() > deadline) {
-      return empty([...recorded.notices, `EAS session list exceeded its total time limit for ${projectRoot}.`]);
-    }
-
-    const parsed = parseSessionListPage(stdout);
-    if (!parsed.ok) return empty([...recorded.notices, parsed.reason]);
-    sessions.push(...parsed.page.sessions);
-    if (!parsed.page.hasNextPage) break;
-
-    const cursor = parsed.page.endCursor?.trim();
-    if (!cursor) {
-      return empty([...recorded.notices, 'EAS session list has a next page but returned no pagination cursor.']);
-    }
-    if (seenCursors.has(cursor)) {
-      return empty([...recorded.notices, `EAS session list repeated pagination cursor ${cursor}.`]);
-    }
-    seenCursors.add(cursor);
-    after = cursor;
-  }
-
-  const lockedRoots = deps.easLockedRoots;
-  if (lockedRoots) {
-    const currentRoots = easLockRoots(projectRoot, loadConfig());
-    if (currentRoots.some((root) => !lockedRoots.includes(root))) {
-      return empty([
-        ...recorded.notices,
-        'EAS session sweep skipped: registered workspace roots changed after remote-session locks were acquired.',
-      ]);
-    }
-  }
-
-  const compared = findOrphanedOwnedSessions({
-    sessions,
-    recordedSessionIds: recorded.ids,
-    projectScope: projectRoot,
-  });
-  const notices = [...recorded.notices, ...compared.notices];
-  if (ledger.notice) notices.push(ledger.notice);
-  const orphaned: ScopedSessionSummary[] = [];
-  if (ledger.safe && recorded.safe) {
-    for (const session of compared.orphaned) {
-      const claim = ledger.claims.get(session.id);
-      if (!claim || claim.name !== session.name || claim.platform !== session.platform) {
-        notices.push(
-          `EAS session ${session.id} (${session.name}) is unclaimed or its fixed ownership record does not match; it was not selected for deletion.`,
-        );
-        continue;
-      }
-      const state = claimStateAbsence(claim);
-      if (state.notice) {
-        notices.push(state.notice);
-        continue;
-      }
-      if (state.absent) orphaned.push(session);
-    }
-  }
-  return {
-    projectScope: projectRoot,
-    orphaned,
-    notices,
-    deletionSafe: recorded.safe && ledger.safe,
-  };
-}
-
-function easLockRoots(projectRoot: string, config: Config | null): string[] {
-  return [...new Set([...Object.keys(config?.projects ?? {}), projectRoot].map(canonicalPath))].toSorted();
-}
-
-function unavailableLockRootNotice(roots: readonly string[]): string | null {
-  for (const root of roots) {
-    try {
-      if (!statSync(root).isDirectory()) throw new Error('path is not a directory');
-      readdirSync(root);
-    } catch (error) {
-      return `EAS session sweep skipped: ${root} is not available for remote-session lock inspection: ${describeError(error)}`;
-    }
-  }
-  return null;
-}
-
-function malformedRemoteSessionLockNotice(roots: readonly string[]): string | null {
-  for (const root of roots) {
-    const lockPath = join(workspaceDir(root), 'remote-session.lock');
-    if (!existsSync(lockPath)) continue;
-    try {
-      const owner = JSON.parse(readFileSync(join(lockPath, 'owner.json'), 'utf-8')) as Record<string, unknown>;
-      if (typeof owner.pid === 'number' && typeof owner.token === 'string') continue;
-    } catch {
-      if (!existsSync(lockPath)) continue;
-    }
-    return `EAS session sweep skipped: the remote-session lock at ${lockPath} has a malformed owner record.`;
-  }
-  return null;
-}
-
-async function withRemoteSessionGcLocks<T>(
-  projectRoot: string,
-  deps: GcDependencies,
-  fn: (coordinatedDeps: GcDependencies) => Promise<T>,
-): Promise<T> {
-  const withLock = deps.withRemoteSessionLock ?? withRemoteSessionLock;
-  const waitMs = deps.easLockWaitMs ?? EAS_LOCK_WAIT_MS;
-  const maxSnapshots = deps.easLockSnapshotRetries ?? EAS_LOCK_SNAPSHOT_RETRIES;
-
-  for (let attempt = 0; attempt < maxSnapshots; attempt++) {
-    const roots = easLockRoots(projectRoot, loadConfig());
-    const unavailable = unavailableLockRootNotice(roots);
-    if (unavailable) return fn({ ...deps, easSweepBlockedNotice: unavailable });
-    const malformed = malformedRemoteSessionLockNotice(roots);
-    if (malformed) return fn({ ...deps, easSweepBlockedNotice: malformed });
-
-    let expanded = false;
-    let coreStarted = false;
-    const acquire = (index: number): Promise<T | null> =>
-      index === roots.length
-        ? (async () => {
-            const rescannedRoots = easLockRoots(projectRoot, loadConfig());
-            if (rescannedRoots.some((root) => !roots.includes(root))) {
-              expanded = true;
-              return null;
-            }
-            coreStarted = true;
-            return fn({ ...deps, easSweepBlockedNotice: undefined, easLockedRoots: roots });
-          })()
-        : withLock(roots[index]!, () => acquire(index + 1), { waitMs });
-
-    try {
-      const result = await acquire(0);
-      if (!expanded) return result as T;
-    } catch (error) {
-      if (coreStarted) throw error;
-      return fn({
-        ...deps,
-        easSweepBlockedNotice: `EAS session sweep skipped: remote-session lock acquisition failed: ${describeError(error)}`,
-      });
-    }
-  }
-
-  return fn({
-    ...deps,
-    easSweepBlockedNotice: `EAS session sweep skipped: registered workspace roots kept changing during remote-session lock acquisition.`,
-  });
-}
-
-function shortKey(key: unknown) {
-  const text = String(key ?? '');
-  return text.length > 6 ? `${text.slice(0, 6)}..` : text;
-}
-
-function parkedAge(parkedAt: string, now: number): string {
-  const at = Date.parse(parkedAt);
-  if (!Number.isFinite(at)) return 'parked at an unknown time';
-  return `parked ${formatLongDuration(Math.max(0, now - at))} ago`;
-}
-
-function formatParkedSimReport(parkedSims: readonly ParkedSimReport[], now: number): string[] {
-  if (parkedSims.length === 0) return [];
-  const known = parkedSims.filter((sim) => sim.bytes !== null);
-  const total = known.reduce((sum, sim) => sum + (sim.bytes ?? 0), 0);
-  const size = known.length === parkedSims.length ? `, ${formatBytes(total)}` : '';
-  const lines = [`Parked simulators (${parkedSims.length}${size}):`];
-  for (const sim of parkedSims) {
-    const model = [sim.model, sim.runtime].filter(Boolean).join(' ');
-    const age = parkedAge(sim.parkedAt, now);
-    const bytes = sim.bytes === null ? '' : ` ${formatBytes(sim.bytes)}`;
-    const gone =
-      sim.listed === false ? ' - not on this machine' : sim.listed === null ? ' - listing unavailable; kept' : '';
-    lines.push(`  ios ${sim.name} (${shortUdid(sim.udid)})${model ? ` ${model}` : ''} ${age}${bytes}${gone}`);
-  }
-  lines.push('              --delete attempts verified deletions and keeps failures.');
-  return lines;
-}
-
-export function formatGcReport(
-  {
-    skipped = [],
-    deadProjects = [],
-    parkedSims = [],
-    orphanedDevices = [],
-    staleDevices = [],
-    staleDeviceRecords = [],
-    buildLocks = { stale: [], live: [] },
-    buildSlots = { stale: [], live: [] },
-    deviceLeases = { expired: [], kept: [] },
-    deviceSweepNotices = [],
-    easSessionSweep = { projectScope: null, orphaned: [], notices: [], deletionSafe: true },
-    caches = [],
-    cacheScope = null,
-    olderThan = null,
-  }: Partial<GcReport>,
-  { now = Date.now() }: { now?: number } = {},
-): string[] {
-  const lines: string[] = [];
-  const staleLocks = buildLocks?.stale ?? [];
-  const liveLocks = buildLocks?.live ?? [];
-  const staleSlots = buildSlots?.stale ?? [];
-  const expiredLeases = deviceLeases?.expired ?? [];
-
-  if (cacheScope) {
-    lines.push(`Cache scope: "${cacheScope}". Devices, project entries and locks were not inspected.`);
-  } else if (
-    deadProjects.length === 0 &&
-    parkedSims.length === 0 &&
-    orphanedDevices.length === 0 &&
-    staleDevices.length === 0 &&
-    staleDeviceRecords.length === 0 &&
-    staleLocks.length === 0 &&
-    staleSlots.length === 0 &&
-    expiredLeases.length === 0 &&
-    easSessionSweep.orphaned.length === 0
-  ) {
-    const reasons = [];
-    if (skipped.length > 0) {
-      reasons.push(`${skipped.length} entr${skipped.length === 1 ? 'y' : 'ies'} could not be checked`);
-    }
-    if (deviceSweepNotices.length > 0) {
-      reasons.push('device sweep incomplete');
-    }
-    if (easSessionSweep.notices.length > 0) {
-      reasons.push('EAS session sweep incomplete');
-    }
-    if (reasons.length > 0) {
-      lines.push(`Nothing to reclaim (${reasons.join('; ')}; see below).`);
-    } else {
-      lines.push('Nothing to reclaim.');
-    }
-  }
-
-  if (deadProjects.length) {
-    lines.push(`Dead project entries (${deadProjects.length}):`);
-    for (const path of deadProjects) lines.push(`  ${path}`);
-  }
-
-  lines.push(...formatParkedSimReport(parkedSims, now));
-
-  if (orphanedDevices.length) {
-    lines.push(`Orphaned devices (${orphanedDevices.length}):`);
-    for (const d of orphanedDevices) lines.push(`  ${d.kind} ${d.name} (${d.id})${deviceSizeSuffix(d)}`);
-  }
-
-  if (staleDevices.length) {
-    lines.push(`Stale owned devices (${staleDevices.length}) - project untouched for ${olderThan ?? '?'}d or more:`);
-    for (const d of staleDevices) {
-      lines.push(`  ${d.kind} ${d.name} (${d.id})${deviceSizeSuffix(d)}`);
-      lines.push(`              ${d.project} (idle ${d.idleDays}d)`);
-    }
-  }
-
-  if (staleDeviceRecords.length) {
-    lines.push(`Stale device records (${staleDeviceRecords.length}) - the device is gone, the project is not:`);
-    for (const r of staleDeviceRecords) {
-      lines.push(`  ${r.kind} ${r.id} is not on this machine`);
-      lines.push(`              recorded by ${r.project}`);
-    }
-    lines.push('              --delete clears the RECORD only; there is no device left to touch.');
-  }
-
-  if (easSessionSweep.orphaned.length) {
-    lines.push(`Orphaned EAS sessions (${easSessionSweep.orphaned.length}) - current EAS project only:`);
-    for (const session of easSessionSweep.orphaned) {
-      const details = [session.platform, session.status].filter(Boolean).join(', ');
-      lines.push(`  ${session.name} (${session.id})${details ? ` [${details}]` : ''}`);
-      lines.push(`              project scope: ${session.projectScope}`);
-      lines.push(`              remedy: eas simulator:stop --id ${session.id}`);
-    }
-  }
-
-  if (staleLocks.length) {
-    lines.push(`Stale build locks (${staleLocks.length}) - the process that was building is gone:`);
-    for (const lock of staleLocks) {
-      lines.push(`  ${lock.platform} ${shortKey(lock.key)} (pid ${lock.pid ?? '?'} is not running)`);
-      lines.push(`              started by ${lock.projectRoot || 'an unrecorded workspace'}`);
-    }
-  }
-
-  if (staleSlots.length) {
-    lines.push(`Stale build slots (${staleSlots.length}) - the process that was building is gone:`);
-    for (const slot of staleSlots) {
-      lines.push(`  slot ${slot.index ?? '?'} (pid ${slot.pid ?? '?'} is not running)`);
-      lines.push(`              held by ${slot.projectRoot || 'an unrecorded workspace'}`);
-    }
-  }
-
-  if (liveLocks.length) {
-    lines.push(`Builds in progress (${liveLocks.length}) - NOT touched, by anything:`);
-    for (const lock of liveLocks) {
-      lines.push(`  ${lock.platform} ${shortKey(lock.key)} (pid ${lock.pid})`);
-      lines.push(`              building in ${lock.projectRoot || 'an unrecorded workspace'}`);
-    }
-  }
-
-  lines.push(...leaseGarbageLines(deviceLeases));
-
-  if (deviceSweepNotices.length) {
-    lines.push(`Device sweep notices (${deviceSweepNotices.length}):`);
-    for (const notice of deviceSweepNotices) lines.push(`  ${notice}`);
-  }
-
-  if (easSessionSweep.notices.length) {
-    lines.push(`EAS session sweep notices (${easSessionSweep.notices.length}):`);
-    for (const notice of easSessionSweep.notices) lines.push(`  ${notice}`);
-  }
-
-  if (skipped.length) {
-    lines.push(`Skipped (${skipped.length}) - not classified as dead:`);
-    for (const entry of skipped) lines.push(`  ${entry.dir}: ${entry.reason}`);
-  }
-
-  if (caches.length) {
-    const total = caches.reduce((n, c) => n + (c.bytes ?? 0), 0);
-    lines.push(`Shared build caches (${caches.length}) - alive, not garbage:`);
-    for (const c of caches) {
-      const tag = c.source ? ` (${c.source})` : '';
-      lines.push(`  ${formatBytes(c.bytes ?? 0).padStart(10)}  ${c.name}${tag}`);
-      lines.push(`              ${c.dir}`);
-      if (c.note) lines.push(`              ${c.note}`);
-      if (c.willEmpty) lines.push('              would be EMPTIED');
-      else if (c.emptySkipped) lines.push(`              would be left alone: ${c.emptySkipped}`);
-    }
-    lines.push(`  total: ${formatBytes(total)}`);
-    const doomed = caches.filter((c) => c.willEmpty);
-    if (doomed.length) {
-      const doomedBytes = doomed.reduce((n, c) => n + (c.bytes ?? 0), 0);
-      lines.push(`  would empty ${doomed.length} of these (${formatBytes(doomedBytes)})`);
-    }
-  }
-
-  return lines;
-}
-
-function leaseGarbageLines({ expired, kept }: DeviceLeaseGarbage): string[] {
-  const lines: string[] = [];
-  if (expired.length) {
-    lines.push(`Expired device leases (${expired.length}) - the device is already free:`);
-    for (const entry of expired) {
-      const name = entry.lease?.deviceName ? ` (${entry.lease.deviceName})` : '';
-      lines.push(`  ${entry.platform} ${entry.id ?? entry.name}${name}`);
-      lines.push(
-        `              held by ${entry.lease?.holder ?? 'an unrecorded workspace'} until ${entry.lease?.expiresAt}`,
-      );
-    }
-    lines.push('              --delete removes the FILE only; an expired lease already holds nothing.');
-  }
-  if (kept.length) {
-    lines.push(`Device lease files kept (${kept.length}) - reported, never deleted:`);
-    for (const entry of kept) lines.push(`  ${entry.name}: ${entry.reason}`);
-  }
-  return lines;
-}
-
-function deviceSizeSuffix(device: { kind: 'ios' | 'android'; bytes?: number }): string {
-  return device.kind === 'android' && device.bytes !== undefined ? ` - ${formatBytes(device.bytes)} on disk` : '';
-}
-
-function withAndroidAvdSizes<T extends { kind: 'ios' | 'android'; id: string; bytes?: number }>(
-  devices: T[],
-  {
-    avdDirectory = ownedAvdDirectory,
-    size = directorySize,
-  }: { avdDirectory?: typeof ownedAvdDirectory; size?: typeof directorySize } = {},
-): T[] {
-  return devices.map((device) => {
-    if (device.kind !== 'android') return device;
-    const dir = avdDirectory(device.id);
-    if (!dir) return device;
-    try {
-      const bytes = size(dir, { timeoutMs: AVD_SIZE_TIMEOUT_MS });
-      return bytes > 0 ? { ...device, bytes } : device;
-    } catch {
-      return device;
-    }
-  });
-}
-
-export function describeParkedSims(
-  records: readonly ParkedSim[],
-  sims: readonly IosSimRecord[],
-  deviceTypes: readonly { identifier: string; name: string }[],
-  { simsChecked = true }: { simsChecked?: boolean } = {},
-): ParkedSimReport[] {
-  const listed = new Map(sims.map((sim) => [sim.udid, sim]));
-  return records.map((record) => {
-    const sim = listed.get(record.udid);
-    return {
-      udid: record.udid,
-      name: sim?.name ?? record.name,
-      model: deviceTypes.find((d) => d.identifier === record.deviceTypeIdentifier)?.name ?? null,
-      runtime: sim ? parseRuntimeVersion(sim.runtime) : parseRuntimeVersion(record.runtimeIdentifier),
-      parkedAt: record.parkedAt,
-      bytes: typeof sim?.dataPathSize === 'number' ? sim.dataPathSize : null,
-      listed: simsChecked ? Boolean(sim) : null,
-    };
-  });
-}
-
-function collectParkedSims(deps: GcDependencies): ParkedSimReport[] {
-  const records = readParked('ios');
-  if (records.length === 0) return [];
-  let sims: IosSimRecord[];
-  let deviceTypes: { identifier: string; name: string }[] = [];
-  try {
-    sims = (deps.listAllIosSims ?? listAllIosSims)({ timeoutMs: DEVICE_LIST_TIMEOUT_MS, includeUnavailable: true });
-  } catch {
-    return describeParkedSims(records, [], [], { simsChecked: false });
-  }
-  try {
-    deviceTypes = (deps.listIosDeviceTypes ?? listIosDeviceTypes)();
-  } catch {}
-  return describeParkedSims(records, sims, deviceTypes);
-}
-
-export function deleteParkedSims(parkedSims: readonly ParkedSimReport[], deps: GcDependencies = {}): number {
-  let failures = 0;
-  let emptied = 0;
-  for (const sim of parkedSims) {
-    if (sim.listed === null) {
-      failures++;
-      console.log(chalk.red(`Could not verify parked ios sim ${sim.name} (${sim.udid}); its pool record was kept.`));
-      continue;
-    }
-    try {
-      const teardown = sim.listed
-        ? teardownParkedIosSim(sim.udid, { label: sim.name, deleteSim: deps.deleteParkedIosSim })
-        : null;
-      const removed = sim.listed ? teardown?.status === 'torn-down' : dropParked('ios', sim.udid);
-      if (!removed) {
-        if (teardown?.status === 'failed') {
-          failures++;
-          console.log(chalk.red(`Failed to delete parked ios sim ${sim.name}: ${teardown.reason}`));
-        } else {
-          console.log(chalk.dim(`Skipped ${sim.name} (${sim.udid}); it is no longer parked.`));
-        }
-        continue;
-      }
-      emptied++;
-      console.log(
-        chalk.green(
-          sim.listed
-            ? `Deleted parked ios sim ${sim.name} (${sim.udid})`
-            : `Dropped the parked record for ${sim.name} (${sim.udid}); it is not on this machine.`,
-        ),
-      );
-    } catch (err) {
-      failures++;
-      console.log(chalk.red(`Failed to delete parked ios sim ${sim.name}: ${(err as Error)?.message || err}`));
-    }
-  }
-  if (emptied) console.log(chalk.dim(`  emptied ${plural(emptied, 'parked simulator')} from the pool`));
-  return failures;
 }
 
 function projectLastTouched(path: string): number {
@@ -943,140 +80,6 @@ function projectLastTouched(path: string): number {
   } catch {
     return NaN;
   }
-}
-
-function deviceSweepIsScoped(unsafeAllowScopedDeviceSweep?: boolean) {
-  return Boolean(process.env.STIM_HOME) && !unsafeAllowScopedDeviceSweep;
-}
-
-function cacheSweepIsScoped() {
-  return Boolean(process.env.STIM_HOME);
-}
-
-function canonicalPath(p: string): string {
-  try {
-    return realpathSync(p);
-  } catch {
-    return resolve(p);
-  }
-}
-
-function isInsideConfigDir(dir: string) {
-  const root = canonicalPath(getConfigDir());
-  const target = canonicalPath(dir);
-  const rel = relative(root, target);
-  return rel !== '' && !rel.startsWith('..') && !isAbsolute(rel);
-}
-
-function machineGlobalReason(cache: CacheDescriptor): string | null {
-  if (!cacheSweepIsScoped()) return null;
-  if (cache.source === 'registered') return null;
-  if (isInsideConfigDir(cache.dir)) return null;
-  return `STIM_HOME scopes this config, but ${cache.dir} is outside it and therefore machine-global`;
-}
-
-const EVERY_CACHE = 'all';
-
-export function selectCaches(caches: CacheDescriptor[], name: string | null | undefined): CacheDescriptor[] {
-  if (!name) return caches;
-  const wanted = name.trim().toLowerCase();
-  if (wanted === EVERY_CACHE) return caches;
-  return caches.filter((c) => c.name.toLowerCase().includes(wanted) || c.dir.toLowerCase().includes(wanted));
-}
-
-function planCacheEmptying(caches: CacheDescriptor[], all: boolean): GcCache[] {
-  const annotated = caches.map((c) => Object.assign({}, c, { machineGlobal: machineGlobalReason(c) }));
-  if (!all) return annotated;
-  return annotated.map((c) => {
-    if (c.prune === 'report-only') {
-      return Object.assign({}, c, {
-        willEmpty: false,
-        emptySkipped: 'report-only shared cache; Stim never deletes it',
-      });
-    }
-    if (c.machineGlobal) {
-      return Object.assign({}, c, { willEmpty: false, emptySkipped: c.machineGlobal });
-    }
-    if (!ownsItsDirectory(c)) {
-      return Object.assign({}, c, {
-        willEmpty: false,
-        emptySkipped: `${c.dir} is not a directory this cache owns`,
-      });
-    }
-    return Object.assign({}, c, { willEmpty: true, emptySkipped: null });
-  });
-}
-
-// A last fail-closed check on the one operation in `gc` that cannot be undone.
-// A cache that carries an explicit `files` list does not own its directory --
-// Metro's maps live loose in os.tmpdir() alongside everything else's temp
-// files -- and a mis-registered `dir` naming a volume root, the user's home,
-// the temp dir or the config dir is a typo, never a cache.
-function ownsItsDirectory(cache: CacheDescriptor): boolean {
-  if (Array.isArray(cache.files)) return false;
-  const dir = canonicalPath(cache.dir);
-  if (dirname(dir) === dir) return false;
-  return ![homedir(), tmpdir(), getConfigDir()].map(canonicalPath).includes(dir);
-}
-
-function emptyCache(cache: CacheDescriptor): {
-  removed: number;
-  bytes: number;
-  skipped: string | null;
-  failed?: number;
-} {
-  if (cache.prune === 'report-only') {
-    return { removed: 0, bytes: 0, skipped: 'report-only shared cache; Stim never deletes it' };
-  }
-  if (cache.prune !== 'atomic') {
-    return pruneCache(cache, { olderThanDays: 0, now: Date.now() + DAY_MS });
-  }
-
-  let names: string[];
-  try {
-    names = readdirSync(cache.dir);
-  } catch (e) {
-    return { removed: 0, bytes: 0, skipped: `could not read ${cache.dir}: ${(e as Error).message}` };
-  }
-  let removed = 0;
-  let failed = 0;
-  for (const name of names) {
-    try {
-      rmSync(join(cache.dir, name), { recursive: true, force: true });
-      removed++;
-    } catch {
-      failed++;
-    }
-  }
-  return { removed, bytes: failed ? 0 : (cache.bytes ?? 0), failed, skipped: null };
-}
-
-function collectDeviceLeases(now: number): DeviceLeaseGarbage {
-  const expired: LeaseFileEntry[] = [];
-  const kept: KeptLeaseFile[] = [];
-  for (const entry of listLeaseFiles()) {
-    const lease = entry.lease;
-    if (!lease) {
-      kept.push({
-        name: entry.name,
-        path: entry.path,
-        reason: 'it does not parse as a lease, so no run may take that device',
-      });
-      continue;
-    }
-    if (leaseIsExpired(lease, now)) {
-      expired.push(entry);
-      continue;
-    }
-    if (!existsSync(lease.holder)) {
-      kept.push({
-        name: entry.name,
-        path: entry.path,
-        reason: `${lease.holder} is not on this machine, but the lease runs until ${lease.expiresAt}`,
-      });
-    }
-  }
-  return { expired, kept };
 }
 
 export async function collectGcReport(
@@ -1415,39 +418,7 @@ async function runGcCore(opts: RunGcOptions, deps: GcDependencies): Promise<void
     }
   }
 
-  function reap(d: OrphanedDevice | StaleProjectDevice) {
-    const r =
-      d.kind === 'ios'
-        ? teardownOwnedIosSim(d.id, { del: true, label: d.name })
-        : teardownOwnedAvd(d.name, { del: true });
-    const what = d.kind === 'ios' ? `ios sim ${d.name} (${d.id})` : `android avd ${d.name}`;
-    if (r.status === 'torn-down') {
-      console.log(chalk.green(`Deleted ${what}`));
-    } else if (r.status === 'missing') {
-      console.log(chalk.dim(`${what} is already gone; nothing to delete.`));
-    } else if (r.status === 'skipped') {
-      console.log(chalk.yellow(`Skipped ${what}: ${r.reason} -- left for a later gc`));
-    } else {
-      deleteFailures++;
-      console.log(chalk.red(`Failed to delete ${d.kind} device ${d.name}: ${r.reason}`));
-    }
-    return r.status;
-  }
-
-  for (const d of orphanedDevices) reap(d);
-
-  for (const d of staleDevices) {
-    const status = reap(d);
-    if (status === 'torn-down' || status === 'missing') {
-      clearDevice(d.project, d.kind);
-      console.log(chalk.dim(`  cleared the ${d.kind} record for ${d.project}`));
-    }
-  }
-
-  for (const r of staleDeviceRecords) {
-    clearDevice(r.project, r.kind);
-    console.log(chalk.green(`Cleared the ${r.kind} record for ${r.project} (${r.id} is not on this machine)`));
-  }
+  deleteFailures += deleteProjectDevices(orphanedDevices, staleDevices, staleDeviceRecords);
 
   for (const lock of buildLocks.stale) {
     const current = readBuildLock(lock.path);
@@ -1497,148 +468,7 @@ async function runGcCore(opts: RunGcOptions, deps: GcDependencies): Promise<void
       console.log(chalk.red(`Failed to clear the device lease at ${entry.path}: ${(err as Error)?.message || err}`));
     }
   }
-  if (easSessionSweep.orphaned.length && easSessionSweep.deletionSafe && easSessionSweep.projectScope) {
-    const projectScope = easSessionSweep.projectScope;
-    const withProjectLock = deps.withEasProjectLock ?? withEasProjectLock;
-    let finalSweepStarted = false;
-    try {
-      await withProjectLock(
-        projectScope,
-        () => {
-          finalSweepStarted = true;
-          return withRemoteSessionGcLocks(projectScope, deps, async (lockedDeps) => {
-            if (lockedDeps.easSweepBlockedNotice) {
-              deleteFailures += easSessionSweep.orphaned.length;
-              console.log(chalk.red(`EAS session deletion refused: ${lockedDeps.easSweepBlockedNotice}`));
-              return;
-            }
-            const coordinatedDeps = lockedDeps;
-            const currentConfig = loadConfig();
-            const currentRoots = easLockRoots(projectScope, currentConfig);
-            const registryExpanded =
-              coordinatedDeps.easLockedRoots &&
-              currentRoots.some((root) => !coordinatedDeps.easLockedRoots!.includes(root));
-            const currentRecords = registryExpanded
-              ? null
-              : readRecordedRemoteSessionIds([...Object.keys(currentConfig?.projects ?? {}), projectScope]);
-            const ledgerRoot = coordinatedDeps.easLedgerRoot ?? easMachineStateRoot();
-            const ledger = readEasSessionLedger(ledgerRoot);
-            const removeClaim = coordinatedDeps.removeEasSessionClaim ?? removeEasSessionClaim;
-            const reconcileClaim = (sessionId: string): void => {
-              let detail = 'the claim store did not remove it';
-              try {
-                if (removeClaim(sessionId, ledgerRoot)) return;
-              } catch (error) {
-                detail = describeError(error);
-              }
-              deleteFailures++;
-              console.log(
-                chalk.red(
-                  `EAS session ${sessionId} is resolved, but its ownership claim could not be removed: ${detail}. The claim was retained for reconciliation.`,
-                ),
-              );
-            };
-            if (registryExpanded) {
-              deleteFailures += easSessionSweep.orphaned.length;
-              console.log(
-                chalk.red('EAS session deletion refused: registered workspace roots changed after classification.'),
-              );
-            } else if (!currentRecords?.safe || !ledger.safe) {
-              deleteFailures += easSessionSweep.orphaned.length;
-              for (const notice of [...(currentRecords?.notices ?? []), ...(ledger.notice ? [ledger.notice] : [])]) {
-                console.log(chalk.red(`EAS session deletion refused: ${notice}`));
-              }
-            } else {
-              const recorded = new Set(currentRecords.ids);
-              const eas = (coordinatedDeps.resolveEasCliBin ?? resolveEasCliBin)(projectScope);
-              const run =
-                coordinatedDeps.runEasFile ?? ((file, args, options) => getExecutor().runFile(file, args, options));
-              for (const session of easSessionSweep.orphaned) {
-                if (recorded.has(session.id)) {
-                  console.log(chalk.dim(`EAS session ${session.id} has a workspace record and was left running.`));
-                  continue;
-                }
-                const claim = ledger.claims.get(session.id);
-                const claimState = claim ? claimStateAbsence(claim) : null;
-                if (
-                  !claim ||
-                  claim.name !== session.name ||
-                  claim.platform !== session.platform ||
-                  !claimState?.absent ||
-                  claimState.notice
-                ) {
-                  deleteFailures++;
-                  console.log(
-                    chalk.red(
-                      `Could not verify fixed ownership for EAS session ${session.id}: ${claimState?.notice ?? 'the claim is missing, mismatched, or still recorded in workspace state'}.`,
-                    ),
-                  );
-                  continue;
-                }
-                if (!eas) {
-                  deleteFailures++;
-                  console.log(chalk.red(`Could not verify EAS session ${session.id}: eas-cli is not available.`));
-                  continue;
-                }
-                const options = {
-                  cwd: session.projectScope,
-                  timeoutMs: EAS_OPERATION_TIMEOUT_MS,
-                  omitEnv: PROXY_CREDENTIAL_ENV,
-                };
-                let live: string;
-                try {
-                  live = run(eas.file, getSessionArgs(session.id), options);
-                } catch (error) {
-                  if (isDefinitiveMissingSessionError(error, session.id)) {
-                    console.log(chalk.dim(`EAS session ${session.id} is already gone.`));
-                    reconcileClaim(session.id);
-                  } else {
-                    deleteFailures++;
-                    console.log(chalk.red(`Could not verify EAS session ${session.id}: ${describeError(error)}`));
-                  }
-                  continue;
-                }
-
-                const inspection = inspectSessionForTeardown(live, session.id);
-                if (inspection.action === 'already-stopped') {
-                  console.log(chalk.dim(`EAS session ${session.id} is already stopped (${inspection.status}).`));
-                  reconcileClaim(session.id);
-                  continue;
-                }
-                if (inspection.action === 'refused') {
-                  deleteFailures++;
-                  console.log(chalk.red(`Could not stop EAS session ${session.id}: ${inspection.reason}`));
-                  continue;
-                }
-
-                let stopped: string;
-                try {
-                  stopped = run(eas.file, stopSessionArgs(session.id), options);
-                } catch (error) {
-                  deleteFailures++;
-                  console.log(chalk.red(`Could not stop EAS session ${session.id}: ${describeError(error)}`));
-                  continue;
-                }
-                const verified = verifyStoppedSession(stopped, session.id);
-                if (!verified.ok) {
-                  deleteFailures++;
-                  console.log(chalk.red(`Could not verify EAS session ${session.id} stopped: ${verified.reason}`));
-                  continue;
-                }
-                console.log(chalk.green(`Stopped EAS session ${session.id} (${session.name})`));
-                reconcileClaim(session.id);
-              }
-            }
-          });
-        },
-        { waitMs: 0, ownerPurpose: 'EAS orphan deletion', machineRoot: deps.easLedgerRoot },
-      );
-    } catch (error) {
-      deleteFailures += easSessionSweep.orphaned.length;
-      const phase = finalSweepStarted ? 'failed' : 'lock acquisition failed';
-      console.log(chalk.red(`EAS session deletion ${phase}: ${describeError(error)}`));
-    }
-  }
+  deleteFailures += await deleteEasSessions(easSessionSweep, deps);
 
   if (deleteFailures) {
     console.log(
@@ -1660,63 +490,7 @@ async function runGcCore(opts: RunGcOptions, deps: GcDependencies): Promise<void
     return;
   }
 
-  let cacheBytes = 0;
-  for (const c of caches) {
-    if (c.machineGlobal) {
-      console.log(chalk.yellow(`Left ${c.name} alone: ${c.machineGlobal}`));
-      continue;
-    }
-    const r = pruneCache(c, { olderThanDays: olderThan });
-    if (r.skipped) {
-      console.log(chalk.yellow(`Left ${c.name} alone: ${r.skipped}`));
-    } else if (r.removed) {
-      cacheBytes += r.bytes;
-      console.log(
-        chalk.green(`Trimmed ${c.name}: ${r.removed} entr${r.removed === 1 ? 'y' : 'ies'} (${formatBytes(r.bytes)})`),
-      );
-    } else {
-      console.log(chalk.dim(`${c.name}: nothing older than ${olderThan}d`));
-    }
-  }
-
-  if (cacheBytes) {
-    console.log(
-      chalk.dim(
-        `Trimmed ${formatBytes(cacheBytes)} of shared cache. The next build that wanted those entries pays to rebuild them.`,
-      ),
-    );
-  }
-}
-
-function emptyCaches(caches: GcCache[]) {
-  let cacheBytes = 0;
-  for (const c of caches) {
-    if (!c.willEmpty) {
-      console.log(chalk.yellow(`Left ${c.name} alone: ${c.emptySkipped}`));
-      continue;
-    }
-    const r = emptyCache(c);
-    if (r.skipped) {
-      console.log(chalk.yellow(`Left ${c.name} alone: ${r.skipped}`));
-    } else if (r.removed) {
-      cacheBytes += r.bytes;
-      console.log(
-        chalk.green(`Emptied ${c.name}: ${r.removed} entr${r.removed === 1 ? 'y' : 'ies'} (${formatBytes(r.bytes)})`),
-      );
-    } else {
-      console.log(chalk.dim(`${c.name}: already empty`));
-    }
-    if (r.failed) {
-      console.log(chalk.red(`  ${r.failed} entr${r.failed === 1 ? 'y' : 'ies'} in ${c.dir} could not be removed`));
-    }
-  }
-  if (cacheBytes) {
-    console.log(
-      chalk.dim(
-        `Emptied ${formatBytes(cacheBytes)} of shared cache. Every build that wanted any of it now pays to rebuild it.`,
-      ),
-    );
-  }
+  trimCaches(caches, olderThan);
 }
 
 export default function gcCommand(program: Command): void {
